bootloader/assets, packaging: rename generateasset to genasset
