+++ bootloader/assets/genasset/export_test.go
