+++ bootloader/assets/generate.go
@@ -19,5 +19,5 @@
 
 package assets
 
-//go:generate go run ./generateasset/main.go -name grub.cfg -in ./data/grub.cfg -out ./grub_cfg_asset.go
-//go:generate go run ./generateasset/main.go -name grub-recovery.cfg -in ./data/grub-recovery.cfg -out ./grub_recovery_cfg_asset.go
+//go:generate go run ./genasset/main.go -name grub.cfg -in ./data/grub.cfg -out ./grub_cfg_asset.go
+//go:generate go run ./genasset/main.go -name grub-recovery.cfg -in ./data/grub-recovery.cfg -out ./grub_recovery_cfg_asset.go
